Code optimization

+++ pkg/datapath/types.go
@@ -194,9 +194,9 @@ func getPtrOffset(ptr unsafe.Pointer, size uintptr) unsafe.Pointer {
 }
 
 func memcpy(dest, src unsafe.Pointer, n uint64) unsafe.Pointer {
-	return C.memcpy(dest, src, n)
+	return C.memcpy(dest, C.int(src), C.uint64(n))
 }
 
 func memset(dest unsafe.Pointer, c int32, n uint64) unsafe.Pointer {
-	return C.memset(dest, c, n)
+	return C.memset(dest, C.int(c), C.uint64(n))
 }
